+++ RELEASE-NOTES.md
@@ -1,5 +1,6 @@
 ## 3.0.M1
 
+* #9 Channel pool support in ChannelFactory
 * #14 Allow multiple routing keys for a combination of exhcnage and queue
 * #15 Make RmqConnectionManager injectable
 * #16 Injectable RmqTopologyManager